refactor: remove inline implementation of removeWasPublicLink and import it from the library for better modularity
